# target/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger API pour GMSSC REST API",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger GMSSC REST API",
    "contact" : {
      "name" : "Bossoh Michel",
      "email" : "michel.bossoh@outlook.fr"
    }
  },
  "basePath" : "http://localhost:8183",
...
  "paths" : {
...
    "/gmssc/api/v1/PieceEquipements/allbyequiment/{IdEquipement}" : {
      "get" : {
        "tags" : [ "gmsscapiv1PieceEquipements" ],
        "operationId" : "listOfPiecEqpmtByEqpmntID",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "IdEquipement",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PieceEquipementDto"
              }
            }
          }
        }
      }
    },
    "/gmssc/api/v1/PieceEquipements/allbypieceequiment/{IdPiece}" : {
      "get" : {
        "tags" : [ "gmsscapiv1PieceEquipements" ],
        "operationId" : "listOfPiecEqpmtByPiece",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "IdPiece",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PieceEquipementDto"
              }
            }
          }
        }
      }
    },
    "/gmssc/api/v1/PieceEquipements/delete/{IdPieceEqpt}" : {
      "delete" : {
        "tags" : [ "gmsscapiv1PieceEquipements" ],
        "operationId" : "deletePieceEquipementById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "IdPieceEqpt",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "/gmssc/api/v1/PieceEquipements/savePieceEquipement" : {
      "post" : {
        "tags" : [ "gmsscapiv1PieceEquipements" ],
        "operationId" : "savePieceEquipement",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PieceEquipementDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PieceEquipementDto"
            }
          }
        }
      }
    },
    "/gmssc/api/v1/PieceEquipements/{IdPieceEqpt}" : {
      "get" : {
        "tags" : [ "gmsscapiv1PieceEquipements" ],
        "operationId" : "getPieceEquipementByID",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "IdPieceEqpt",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PieceEquipementDto"
            }
          }
        }
      }
    },
    "/gmssc/api/v1/UtilisateurGroupeIntervenant/delete/{IduserGrpeInterv}" : {
      "delete" : {
        "tags" : [ "gmsscapiv1utilisateurGroupeIntervenant" ],
        "operationId" : "deleteUtilisateurGroupeIntervenantById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "IduserGrpeInterv",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
...
    "/gmssc/api/v1/actions/delete/{Idaction}" : {
      "delete" : {
        "tags" : [ "gmsscapiv1actions" ],
        "operationId" : "deleteActiobById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "Idaction",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "/gmssc/api/v1/actions/saveaction" : {
      "post" : {
        "tags" : [ "gmsscapiv1actions" ],
        "operationId" : "saveAction",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/TodoMaintenanceDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/TodoMaintenanceDto"
            }
          }
        }
      }
    },
    "/gmssc/api/v1/actions/{Idaction}" : {
      "get" : {
        "tags" : [ "gmsscapiv1actions" ],
        "operationId" : "getActionByID",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "Idaction",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/TodoMaintenanceDto"
            }
          }
        }
      }
    },
...
    "/gmssc/api/v1/utilisateurGroupeIntervenant/allparGroupeIntervenant/{IduserGrpeInterv}" : {
      "get" : {
        "tags" : [ "gmsscapiv1utilisateurGroupeIntervenant" ],
        "operationId" : "getlistOfUtilisateurGroupeIntervenantByGroupeIntervenant",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "IduserGrpeInterv",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/IntervenantGroupeIntervenantDto"
              }
            }
          }
        }
      }
    },
    "/gmssc/api/v1/utilisateurGroupeIntervenant/saveusergroupeintervenant" : {
      "post" : {
        "tags" : [ "gmsscapiv1utilisateurGroupeIntervenant" ],
        "operationId" : "saveUtilisateurGroupeIntervenant",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/IntervenantGroupeIntervenantDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/IntervenantGroupeIntervenantDto"
            }
          }
        }
      }
    },
    "/gmssc/api/v1/utilisateurGroupeIntervenant/{IduserGrpeInterv}" : {
      "get" : {
        "tags" : [ "gmsscapiv1utilisateurGroupeIntervenant" ],
        "operationId" : "getUtilisateurGroupeIntervenantByID",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "IduserGrpeInterv",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/IntervenantGroupeIntervenantDto"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "AdresseDto" : {
      "type" : "object",
      "properties" : {
        "pays" : {
          "type" : "string"
        },
        "ville" : {
          "type" : "string"
        },
        "rue" : {
          "type" : "string"
        },
        "codePostale" : {
          "type" : "string"
        },
        "adresse1" : {
          "type" : "string"
        },
        "adresse2" : {
          "type" : "string"
        }
      }
    },
...
    "BienImmobilierDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeBienImmobilier" : {
          "type" : "string"
        },
        "nomBienImmobilier" : {
          "type" : "string"
        },
        "typeBienImmobilier" : {
          "type" : "string"
        },
        "nbreBatiments" : {
          "type" : "integer",
          "format" : "int32"
        },
        "mobile" : {
          "type" : "string"
        },
        "telephone" : {
          "type" : "string"
        },
        "villeSociete" : {
          "type" : "string"
        },
        "commune" : {
          "type" : "string"
        },
        "quartier" : {
          "type" : "string"
        },
        "autreInformation" : {
          "type" : "string"
        },
        "nombrePiece" : {
          "type" : "integer",
          "format" : "int32"
        },
        "societeDto" : {
          "$ref" : "#/definitions/SocieteDto"
        },
        "adresses1" : {
          "type" : "string"
        },
        "adresses2" : {
          "type" : "string"
        }
      }
    },
...
    "ContratDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeContrat" : {
          "type" : "string"
        },
        "dateSignature" : {
          "type" : "string",
          "format" : "date"
        },
        "dureeContrat" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date"
        },
        "contactUrgence" : {
          "type" : "string"
        },
        "delaiIntervention" : {
          "type" : "string"
        },
        "resiliation" : {
          "type" : "boolean"
        },
        "verificationControleGeneral" : {
          "type" : "boolean"
        },
        "verificationFixationsSupports" : {
          "type" : "boolean"
        },
        "verificationCircuitFrigorifique" : {
          "type" : "boolean"
        },
        "detectionFuitesEventuelles" : {
          "type" : "boolean"
        },
        "verificationBonneMarcheSecurites" : {
          "type" : "boolean"
        },
        "nettoyageFiltre" : {
          "type" : "boolean"
        },
        "depoussierageNettoyageUnites" : {
          "type" : "boolean"
        },
        "verificationEcoulementEaux" : {
          "type" : "boolean"
        },
        "desinfectionNettoyageBacs" : {
          "type" : "boolean"
        },
        "nettoyageEnlevementResidus" : {
          "type" : "boolean"
        },
        "etablissementRapportVisite" : {
          "type" : "boolean"
        },
        "risquesEventuels" : {
          "type" : "boolean"
        },
        "coutInterventionsInclus" : {
          "type" : "boolean"
        },
        "coutInterventionsExclus" : {
          "type" : "boolean"
        },
        "coutInterventionsFacture" : {
          "type" : "boolean"
        },
        "societeDto" : {
          "$ref" : "#/definitions/SocieteDto"
        },
        "equipementDtos" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/EquipementDto"
          }
        }
      }
    },
    "CorrectiveDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "actionsMener" : {
          "type" : "string"
        },
        "reparation" : {
          "type" : "boolean"
        },
        "depannage" : {
          "type" : "boolean"
        },
        "typeTravaux" : {
          "type" : "string"
        },
        "coutMaintenance" : {
          "type" : "number",
          "format" : "double"
        },
        "dateDebutIntervention" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFinIntervention" : {
          "type" : "string",
          "format" : "date"
        },
        "resolu" : {
          "type" : "boolean"
        },
        "rapportInterv" : {
          "type" : "boolean"
        },
        "groupeIntervenantDto" : {
          "$ref" : "#/definitions/GroupeIntervenantDto"
        }
      }
    },
    "EquipementDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeEquipement" : {
          "type" : "string"
        },
        "marque" : {
          "type" : "string"
        },
        "numeroSerie" : {
          "type" : "string"
        },
        "modele" : {
          "type" : "string"
        },
        "hauteur" : {
          "type" : "number",
          "format" : "double"
        },
        "largeur" : {
          "type" : "number",
          "format" : "double"
        },
        "poidsNetInterieur" : {
          "type" : "number",
          "format" : "double"
        },
        "poidsNetExterieur" : {
          "type" : "number",
          "format" : "double"
        },
        "puissance" : {
          "type" : "string"
        },
        "dateAchat" : {
          "type" : "string",
          "format" : "date"
        },
        "garantie" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nomFournisseur" : {
          "type" : "string"
        },
        "personneRessource" : {
          "type" : "string"
        },
        "telephone" : {
          "type" : "string"
        },
        "mobile" : {
          "type" : "string"
        },
        "livreAvecAccessoires" : {
          "type" : "boolean"
        },
        "description" : {
          "type" : "string"
        },
        "numFacture" : {
          "type" : "string"
        },
        "prixFacture" : {
          "type" : "string"
        },
        "dateMiseService" : {
          "type" : "string",
          "format" : "date"
        },
        "etatService" : {
          "type" : "string"
        },
        "typeClim" : {
          "type" : "string"
        },
        "contratDto" : {
          "$ref" : "#/definitions/ContratDto"
        },
        "societeDto" : {
          "$ref" : "#/definitions/SocieteDto"
        },
        "longueur" : {
          "type" : "number",
          "format" : "double"
        }
      }
    },
    "GroupeIntervenantDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "dateAffectation" : {
          "type" : "string",
          "format" : "date"
        },
        "commentaire" : {
          "type" : "string"
        },
        "signalerPanneDto" : {
          "$ref" : "#/definitions/SignalerPanneDto"
        },
        "dateFinTravaux" : {
          "type" : "string",
          "format" : "date"
        }
      }
    },
    "IntervenantDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nomInterv" : {
          "type" : "string"
        },
        "prenomInterv" : {
          "type" : "string"
        },
        "fonctionInterv" : {
          "type" : "string"
        },
        "mobile" : {
          "type" : "string"
        },
        "adresse" : {
          "type" : "string"
        },
        "emailIntervenant" : {
          "type" : "string"
        },
        "autreInfo" : {
          "type" : "string"
        },
        "societeDto" : {
          "$ref" : "#/definitions/SocieteDto"
        }
      }
    },
    "IntervenantGroupeIntervenantDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "dateAffectionUser" : {
          "type" : "string",
          "format" : "date"
        },
        "groupeIntervenantDto" : {
          "$ref" : "#/definitions/GroupeIntervenantDto"
        },
        "intervenantDto" : {
          "$ref" : "#/definitions/IntervenantDto"
        }
      }
    },
    "PieceEquipementDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "dateInstallation" : {
          "type" : "string",
          "format" : "date"
        },
        "equipementDto" : {
          "$ref" : "#/definitions/EquipementDto"
        },
        "piecesDto" : {
          "$ref" : "#/definitions/PiecesDto"
        }
      }
    },
    "PiecesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codePiece" : {
          "type" : "string"
        },
        "nomPiece" : {
          "type" : "string"
        },
        "nomBatiment" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "capacitePiece" : {
          "type" : "integer",
          "format" : "int32"
        },
        "positionEtage" : {
          "type" : "integer",
          "format" : "int32"
        },
        "typeSalle" : {
          "type" : "string"
        },
        "bienImmobilierDto" : {
          "$ref" : "#/definitions/BienImmobilierDto"
        }
      }
    },
...
    "RolesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "roleName" : {
          "type" : "string"
        }
      }
    },
    "SignalerPanneDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeSignalerPanne" : {
          "type" : "string"
        },
        "objetPanne" : {
          "type" : "string"
        },
        "descriptionPanne" : {
          "type" : "string"
        },
        "dateSignalerPanne" : {
          "type" : "string",
          "format" : "date"
        },
        "priorite" : {
          "type" : "string"
        },
        "dateResolutionPanne" : {
          "type" : "string",
          "format" : "date"
        },
        "dateAnnulation" : {
          "type" : "string",
          "format" : "date"
        },
        "etatSignalerPanne" : {
          "type" : "string"
        },
        "causeAnnulation" : {
          "type" : "string"
        },
        "suggestionsAmeliration" : {
          "type" : "string"
        },
        "utilisateurDto" : {
          "$ref" : "#/definitions/UtilisateurDto"
        },
        "piecesDto" : {
          "$ref" : "#/definitions/PiecesDto"
        },
        "intervenantDto" : {
          "$ref" : "#/definitions/IntervenantDto"
        }
      }
    },
    "SocieteDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeSociete" : {
          "type" : "string"
        },
        "denomination" : {
          "type" : "string"
        },
        "sigle" : {
          "type" : "string"
        },
        "dateCreationSociete" : {
          "type" : "string",
          "format" : "date"
        },
        "descriptionActivite" : {
          "type" : "string"
        },
        "codeFiscal" : {
          "type" : "string"
        },
        "photo" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "mobile" : {
          "type" : "string"
        },
        "fax" : {
          "type" : "string"
        },
        "siteWeb" : {
          "type" : "string"
        },
        "societeMaintenance" : {
          "type" : "boolean"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        }
      }
    },
    "TodoMaintenanceDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "titre" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "quantite" : {
          "type" : "integer",
          "format" : "int32"
        },
        "unite" : {
          "type" : "string"
        },
        "dateDoto" : {
          "type" : "string",
          "format" : "date"
        },
        "lieuGeographique" : {
          "type" : "string"
        },
        "coutMaintenance" : {
          "type" : "number",
          "format" : "double"
        },
        "correctiveDto" : {
          "$ref" : "#/definitions/CorrectiveDto"
        }
      }
    },
    "UtilisateurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "mobile" : {
          "type" : "string"
        },
        "dateDeNaissance" : {
          "type" : "string",
          "format" : "date"
        },
        "moteDePasse" : {
          "type" : "string"
        },
        "adresseDto" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "photo" : {
          "type" : "string"
        },
        "roles" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/RolesDto"
          }
        }
      }
    }
  }
}